--- lesson4/engine/pkg/crawler/spider/go.mod
@@ -1,7 +1,0 @@
-module go.core/lesson4/engine/pkg/crawler/spider
-
-go 1.13
-
-require (
-	golang.org/x/net v0.0.0-20201029221708-28c70e62bb1d
-)
