--- lesson4/engine/pkg/crawler/stub/go.mod
@@ -1,3 +1,0 @@
-module go.core/lesson4/engine/pkg/crawler/stub
-
-go 1.13
