--- lesson4/engine/pkg/document/go.mod
@@ -1,3 +1,0 @@
-module go.core/lesson4/engine/pkg/document
-
-go 1.15
