--- lesson4/engine/pkg/index/go.mod
@@ -1,9 +1,0 @@
-module go.core/lesson4/engine/pkg/index
-
-go 1.15
-
-replace go.core/lesson4/engine/pkg/document v0.0.0 => ./../document
-
-require (
-	go.core/lesson4/engine/pkg/document v0.0.0
-)
