+++ lesson4/pkg/crawler/stub/go.mod
@@ -1,0 +1,3 @@
+module go.core/lesson4/pkg/crawler/stub
+
+go 1.13
