+++ lesson4/pkg/document/go.mod
@@ -1,0 +1,3 @@
+module go.core/lesson4/pkg/document
+
+go 1.15
