+++ lesson4/pkg/index/go.mod
@@ -1,0 +1,9 @@
+module go.core/lesson4/pkg/index
+
+go 1.15
+
+replace go.core/lesson4/pkg/document v0.0.0 => ../document
+
+require (
+	go.core/lesson4/pkg/document v0.0.0
+)
