--- lesson5/engine/pkg/crawler/membot/go.mod
@@ -1,3 +1,0 @@
-module go.core/lesson5/engine/pkg/crawler/membot
-
-go 1.13
