+++ lesson5/pkg/crawler/go.mod
@@ -1,0 +1,9 @@
+module go.core/lesson5/pkg/crawler
+
+go 1.13
+
+replace go.core/lesson5/pkg/crawler/membot v0.0.0 => /membot
+
+require (
+	go.core/lesson5/pkg/crawler/membot v0.0.0
+)
