+++ lesson5/pkg/crawler/membot/go.mod
@@ -1,0 +1,3 @@
+module go.core/lesson5/pkg/crawler/membot
+
+go 1.13
